Add .py.[hc]pp extension
