+++ CMakeLists.txt
@@ -43,10 +43,10 @@ find_package(pybind11 CONFIG REQUIRED)
 
 # Find C++ files
 file(REAL_PATH src SOURCE_PATH)
-file(GLOB_RECURSE AMULET_NBT_EXTENSION_SOURCES LIST_DIRECTORIES false ${SOURCE_PATH}/amulet/nbt/pybind/*.cpp)
-file(GLOB_RECURSE AMULET_NBT_EXTENSION_HEADERS LIST_DIRECTORIES false ${SOURCE_PATH}/amulet/nbt/include/amulet_nbt/pybind/*.hpp)
-file(GLOB_RECURSE AMULET_NBT_SOURCES LIST_DIRECTORIES false ${SOURCE_PATH}/amulet/nbt/*.cpp)
-file(GLOB_RECURSE AMULET_NBT_HEADERS LIST_DIRECTORIES false ${SOURCE_PATH}/amulet/nbt/*.hpp)
+file(GLOB_RECURSE AMULET_NBT_EXTENSION_SOURCES LIST_DIRECTORIES false ${SOURCE_PATH}/amulet/*.py.cpp)
+file(GLOB_RECURSE AMULET_NBT_EXTENSION_HEADERS LIST_DIRECTORIES false ${SOURCE_PATH}/amulet/*.py.hpp)
+file(GLOB_RECURSE AMULET_NBT_SOURCES LIST_DIRECTORIES false ${SOURCE_PATH}/amulet/*.cpp)
+file(GLOB_RECURSE AMULET_NBT_HEADERS LIST_DIRECTORIES false ${SOURCE_PATH}/amulet/*.hpp)
 list(REMOVE_ITEM AMULET_NBT_SOURCES ${AMULET_NBT_EXTENSION_SOURCES})
 list(REMOVE_ITEM AMULET_NBT_HEADERS ${AMULET_NBT_EXTENSION_HEADERS})
