+++ src/amulet/nbt/include/amulet_nbt/pybind/encoding.py.hpp
