+++ src/amulet/nbt/pybind/snbt.py.cpp
