+++ src/amulet/nbt/pybind/tag/py_compound_tag.py.cpp
@@ -14,8 +14,8 @@
 #include <amulet_nbt/common.hpp>
 #include <amulet_nbt/nbt_encoding/binary.hpp>
 #include <amulet_nbt/nbt_encoding/string.hpp>
-#include <amulet_nbt/pybind/encoding.hpp>
-#include <amulet_nbt/pybind/serialisation.hpp>
+#include <amulet_nbt/pybind/encoding.py.hpp>
+#include <amulet_nbt/pybind/serialisation.py.hpp>
 #include <amulet_nbt/tag/abc.hpp>
 #include <amulet_nbt/tag/compound.hpp>
 #include <amulet_nbt/tag/copy.hpp>
