+++ src/amulet/nbt/pybind/tag/py_list_tag.py.cpp
@@ -17,8 +17,8 @@
 
 #include <amulet_nbt/nbt_encoding/binary.hpp>
 #include <amulet_nbt/nbt_encoding/string.hpp>
-#include <amulet_nbt/pybind/encoding.hpp>
-#include <amulet_nbt/pybind/serialisation.hpp>
+#include <amulet_nbt/pybind/encoding.py.hpp>
+#include <amulet_nbt/pybind/serialisation.py.hpp>
 #include <amulet_nbt/tag/abc.hpp>
 #include <amulet_nbt/tag/copy.hpp>
 #include <amulet_nbt/tag/eq.hpp>
